+++ backend/server/src/api/public.rs
@@ -1,5 +1,7 @@
 mod course_contents;
+mod defaults;
 mod translations;
 
 pub use course_contents::*;
+pub use defaults::*;
 pub use translations::*;